requirements.txt
# ── Web Framework ─────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# ── Database ──────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# ── Auth & Security ───────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ── Validation & Config ───────────────────────────────────
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
email-validator==2.2.0

# ── HTTP Client ───────────────────────────────────────────
httpx==0.28.1

# ── Document Processing ───────────────────────────────────
beautifulsoup4==4.12.3
pypdf==5.1.0

# ── Payments ──────────────────────────────────────────────
mercadopago==2.3.0

# ── Email ─────────────────────────────────────────────────
aiosmtplib==3.0.2

# ── AI Providers ──────────────────────────────────────────
google-genai==1.70.0
cerebras-cloud-sdk==1.67.0
groq==0.18.0
